requirements.txt
absl-py==0.12.0
alabaster==0.7.12
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
docutils==0.16
entrypoints==0.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.37.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==4.0.0
ipykernel==5.5.3
ipyparallel==6.3.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.4.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nose==1.3.7
notebook==6.3.0
numpy @ file:///D:/bld/numpy_1616933410625/work
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
pycparser==2.20
pydicom==2.1.2
Pygments==2.8.1
pyparsing==2.4.7
pyreadline @ file:///D:/bld/pyreadline_1611175297308/work
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.2
Send2Trash==1.5.0
Shapely==1.7.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
snowballstemmer==2.1.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==1.15.0
tensorflow==1.15.3
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tifffile==2021.4.8
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1

numpy~=1.20.2
pip~=21.0.1
setuptools~=52.0.0